Change: Restructure code
