+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "c4-e5-chess"
-version = "0.2.1"
+version = "0.3.0"
 authors = ["Eugen Lindorfer"]
 edition = "2021"
 description = "C4-E5 Chess is a UCI compatible chess engine based on the move generator in crate Chess. These features are provided: Parallelised iterative depthening, late move pruning, principal variant search, transposition table."
