+++ src/cmd/time_management.rs
@@ -1,5 +1,6 @@
 use super::constants::*;
-use crate::engine::{game::Game, types::*};
+use crate::engine::game::Game;
+use crate::misc::types::*;
 use chess::Color;
 use log::info;
 use std::cmp::{max, min};
