+++ src/engine/mod.rs
@@ -1,14 +1,8 @@
 /// Some useful constants
 pub mod constants;
 
-/// The evaluation function
-pub mod evaluate;
-
 /// A chess game
 pub mod game;
-
-/// Collection of helpers
-pub mod helpers;
 
 /// Position history counter
 pub mod history;
@@ -21,6 +15,3 @@ pub mod pvs;
 
 /// Transposition table
 pub mod store;
-
-/// Type definitions
-pub mod types;
