+++ src/eval/mod.rs
@@ -1,0 +1,8 @@
+/// Some useful constants
+pub mod constants;
+
+/// The evaluation function
+pub mod evaluate;
+
+/// Collection of helpers
+pub mod helpers;
