+++ src/misc/mod.rs
@@ -1,2 +1,5 @@
 /// Collection of helpers
 pub mod helpers;
+
+/// Type definitions
+pub mod types;
